+++ frontend/src/pages/Admin/CouponDesign/ChoiceTemplate/index.tsx
@@ -1,13 +1,13 @@
 import { Dispatch, SetStateAction, useRef, useState } from 'react';
 import { ChoiceTemplateContainer, SampleImg, SampleImageContainer } from './style';
-import TabBar from '../../../../../components/TabBar';
-import { TEMPLATE_MENU, TEMPLATE_OPTIONS } from '../../../../../constants';
+import TabBar from '../../../../components/TabBar';
+import { TEMPLATE_MENU, TEMPLATE_OPTIONS } from '../../../../constants';
 import { useQuery } from '@tanstack/react-query';
 import { useLocation } from 'react-router-dom';
-import { getCouponSamples } from '../../../../../api/get';
-import { parseStampCount } from '../../../../../utils';
-import { SampleBackCouponImage, SampleImage, StampCoordinate } from '../../../../../types';
-import { SampleCouponRes } from '../../../../../types/api';
+import { getCouponSamples } from '../../../../api/get';
+import { parseStampCount } from '../../../../utils';
+import { SampleBackCouponImage, SampleImage, StampCoordinate } from '../../../../types';
+import { SampleCouponRes } from '../../../../types/api';
 
 interface ChoiceTemplateProps {
   frontImage: string;
