+++ frontend/src/pages/Admin/CouponDesign/ChoiceTemplate/style.tsx
